Move tags into the `graph` package
