+++ tests/unit/test_io.py
@@ -33,8 +33,8 @@ import nvtabular as nvt
 import nvtabular.io
 from nvtabular import dispatch, ops
 from nvtabular.graph.schema import Schema
+from nvtabular.graph.tags import Tags
 from nvtabular.io.parquet import GPUParquetWriter
-from nvtabular.tags import Tags
 from tests.conftest import allcols_csv, mycols_csv, mycols_pq, run_in_context
 
 cudf = pytest.importorskip("cudf")
